changed rand48 routines to use c++11 <random> library

+++ gunrock/csr.cuh
@@ -375,7 +375,7 @@ struct Csr
         {
             input_label.read(reinterpret_cast<char*>(node_values), v * sizeof(Value));
         }
-	for(int i=0; i<v; i++) printf("%lld ", node_values[i]); printf("\n");
+	    for(int i=0; i<v; i++) printf("%lld ", (long long)node_values[i]); printf("\n");
 
         time_t mark2 = time(NULL);
         if (!quiet)
